Make the model naming more intuitive (#3)
This makes it easier to just supply a network name for training and prediction and not worry about inconsistency when making predictions, where you used to have to supply the seed name for the weights to be picked up. It was a little thing, but pretty annoying and certainly not something to have in the notebooks.