Dev #20
* initial docstrings for modules and functions
* type hints for function and class parameters
* cleared up some rough elements

This is the start of work towards being more maintainable/clearer code and documentation layout

There are a lot of things I dislike about type hints, but for this project and the audience they
make some sense. I dislike typing within the docstrings, so the happy medium is to use type hints
in the code.

Once we move past 3.9, the type hints can become stronger, at present only simple types and objects
are represented because I dislike the use of "import typing" for anything stronger within 3.8's lineage

This should make the code more accessible in the long run. The introduction of docstrings
is just well overdue, typehinting aside...